+++ src/main/kotlin/com/jfrog/conan/clion/services/ConanService.kt
@@ -3,6 +3,7 @@ package com.jfrog.conan.clion.services
 import com.intellij.ide.util.PropertiesComponent
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
+import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.Messages
 import com.intellij.openapi.vfs.LocalFileSystem
@@ -12,7 +13,7 @@ import com.jfrog.conan.clion.conan.ConanPluginUtils
 import com.jfrog.conan.clion.conan.extensions.downloadFromUrl
 import com.jfrog.conan.clion.models.LibraryData
 import com.jfrog.conan.clion.models.PersistentStorageKeys
-import kotlinx.coroutines.runBlocking
+import kotlinx.serialization.SerializationException
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
 import java.io.File
@@ -159,9 +160,7 @@ class ConanService(val project: Project) {
 
         if (!targetFile.exists() || update && ConanPluginUtils.fileHasOverwriteComment(targetFile)) {
             targetFile.parentFile.mkdirs()
-            runBlocking {
-                targetFile.downloadFromUrl(cmakeProviderURL)
-            }
+            targetFile.downloadFromUrl(cmakeProviderURL)
             // Re-write it, but adding the overwrite header
             ConanPluginUtils.writeToFileWithOverwriteComment(targetFile, targetFile.readText())
             LocalFileSystem.getInstance().refreshAndFindFileByIoFile(targetFile)
@@ -182,20 +181,29 @@ class ConanService(val project: Project) {
 
         if (!targetFile.exists() || update) {
             targetFile.parentFile.mkdirs()
-            runBlocking {
-                targetFile.downloadFromUrl(remoteDataURL)
-            }
+            targetFile.downloadFromUrl(remoteDataURL)
         }
         if (targetFile.exists()) {
             val libraryData = targetFile.readText()
 
-            fireOnLibraryDataChanged(Json{ignoreUnknownKeys=true}.decodeFromString<LibraryData>(libraryData))
+            try {
+                val parsedJson = Json{ignoreUnknownKeys=true}.decodeFromString<LibraryData>(libraryData)
+                fireOnLibraryDataChanged(parsedJson)
+            } catch (e: SerializationException) {
+                thisLogger().error(e)
+                fireOnLibraryDataChanged(LibraryData(hashMapOf()))
+            }
         }
     }
 
     fun getRemoteData(): LibraryData {
-        val targetData = getRemoteDataText()
-        return Json{ignoreUnknownKeys=true}.decodeFromString<LibraryData>(targetData)
+        return try {
+            val targetData = getRemoteDataText()
+            Json { ignoreUnknownKeys = true }.decodeFromString<LibraryData>(targetData)
+        } catch (e: SerializationException) {
+            thisLogger().error(e)
+            LibraryData(hashMapOf())
+        }
     }
 
     fun getRemoteDataText(): String {